fix imports for filter types & permission types
