+++ src/schema/query/recordsAggregation.query.ts
@@ -25,7 +25,7 @@ import { accessibleBy } from '@casl/mongoose';
 import { GraphQLDate } from 'graphql-scalars';
 import { graphQLAuthCheck } from '@schema/shared';
 import { Context } from '@server/apollo/context';
-import { CompositeFilterDescriptor } from '@const/compositeFilter';
+import { CompositeFilterDescriptor } from '../../types/filter';
 
 /** Pagination default items per query */
 const DEFAULT_FIRST = 10;
