+++ src/types/filter.ts
@@ -1,4 +1,41 @@
-import { filterOperator } from './filterOperators';
+/**
+ * Represents the supported operators.
+ *
+ * `"eq"` (equal to)
+ * `"neq"` (not equal to)
+ * `"isnull"` (is equal to null)
+ * `"isnotnull"` (is not equal to null)
+ * `"lt"` (less than)
+ * `"lte"` (less than or equal to)
+ * `"gt"` (greater than)
+ * `"gte"` (greater than or equal to)
+ *
+ * The following operators are supported for string fields only:
+ * `"startswith"`
+ * `"endswith"`
+ * `"contains"`
+ * `"doesnotcontain"`
+ * `"isempty"`
+ * `"isnotempty"`
+ */
+export declare enum filterOperator {
+  CONTAINS = 'contains',
+  DOES_NOT_CONTAIN = 'doesnotcontain',
+  DOES_NOT_END_WITH = 'doesnotendwith',
+  DOES_NOT_START_WITH = 'doesnotstartwith',
+  ENDS_WITH = 'endswith',
+  EQUAL_TO = 'eq',
+  GREATER_THAN = 'gt',
+  GREATER_THAN_OR_EQUAL = 'gte',
+  IS_EMPTY = 'isempty',
+  IS_NOT_EMPTY = 'isnotempty',
+  IS_NOT_NULL = 'isnotnull',
+  IS_NULL = 'isnull',
+  LESS_THAN = 'lt',
+  LESS_THAN_OR_EQUAL = 'lte',
+  NOT_EQUAL_TO = 'neq',
+  STARTS_WITH = 'startswith',
+}
 
 /**
  * Type for the filter descriptor used by @progress/kendo-data-query on frontend.
