+++ src/types/index.ts
@@ -1,1 +1,2 @@
-export * from './permission.type';
+export * from './permission';
+export * from './filter';
