+++ tsconfig.json
@@ -15,9 +15,7 @@
       "@security/*": ["security/*"],
       "@server/*": ["server/*"],
       "@services/*": ["services/*"],
-      "@utils/*": ["utils/*"],
-      "@types": ["types"],
-      "@types/*": ["types/*"]
+      "@utils/*": ["utils/*"]
     },
     "esModuleInterop": true,
     "skipLibCheck": true,
